EE-164: move storage proptests to storage
