+++ execution-engine/shared/src/lib.rs
@@ -2,3 +2,4 @@ extern crate blake2;
 extern crate common;
 
 pub mod newtypes;
+pub mod test_utils;
