+++ execution-engine/storage/Cargo.toml
@@ -11,3 +11,6 @@ common = { path = "../common", features = ["std"], package = "casperlabs-contrac
 parking_lot = "0.7.1"
 shared = { path = "../shared" }
 
+[dev-dependencies]
+gens = { path = "../gens" }
+proptest = "0.9.2"
