+++ execution-engine/tests/src/bytesrepr.rs
@@ -1,20 +1,7 @@
-use common::bytesrepr::*;
 use gens::gens::*;
 use proptest::collection::vec;
 use proptest::prelude::*;
-
-fn test_serialization_roundtrip<T: ToBytes + FromBytes + PartialEq + std::fmt::Debug>(
-    el: &T,
-) -> bool {
-    match deserialize::<T>(&ToBytes::to_bytes(el))
-        .map(|r| r == *el)
-        .ok()
-    {
-        Some(true) => true,
-        Some(false) => false,
-        None => false,
-    }
-}
+use shared::test_utils::test_serialization_roundtrip;
 
 proptest! {
 
@@ -111,26 +98,6 @@ proptest! {
     #[test]
     fn test_value_serialization(v in value_arb()) {
         assert!(test_serialization_roundtrip(&v));
-    }
-
-    #[test]
-    fn test_blake2b_hash(hash in blake2b_hash_arb()) {
-        assert!(test_serialization_roundtrip(&hash));
-    }
-
-    #[test]
-    fn test_trie_pointer(pointer in trie_pointer_arb()) {
-        assert!(test_serialization_roundtrip(&pointer));
-    }
-
-    #[test]
-    fn test_trie_pointer_block(pointer_block in trie_pointer_block_arb()) {
-        assert!(test_serialization_roundtrip(&pointer_block));
-    }
-
-    #[test]
-    fn test_trie(trie in trie_arb()) {
-        assert!(test_serialization_roundtrip(&trie));
     }
 
     #[test]
